isInserted:Add bool isInserted to trigger the toast and provide feedback to user
